add the special mov32 for zext
The extra assumption/invariant provided by the verifier is that an
instruction with BPF_MOV|BPF_X must imply BPF_ALU (32-bit) and imm
== 1.  In other words, the verifier will reject 64-bit BPF_MOV|BPF_X
with a non-zero imm (which is stronger than what we need here).

To test this, set src and dst to the same value - right now they are
different (0 and 1) and the proof would be meaningless.