lean: tweak lean proof to kill one axiom.
The axiom about wf safety is unnecessary. It should (in principle) be
implied by the WF termination axiom, but this wasn't true because the
safety definition required there to be a valid instruction to execute
next even for final states. Weaken the safety to include final states as
well, then rewrite the axiom as a lemma and tweak the star_src_correct
proof to kill the final state case before proceeding with the usual
per-instruction correctness.