Fix rv64 LD_IMM64 verification by delaying inline bv assumptions
The implementation for loading a 64-bit immediate in the RV64 JIT
contains the following sequence of code:

  shift = __ffs(upper);
  upper >>= shift;

We model the result of __ffs as a fresh symbolic value, with
assumptions on its correctness. This commit tweaks the axiomatization
in two ways that causes verification to no longer time out.
This is particularly difficult as reasoning about arithmetic right
shifts by symbolic amounts can be expensive.

First, change the axiomatization from

  (bvult (ffs x) (bv 64 64)))
  (bveq (bvshl (bvlshr x (ffs x)) (ffs x)) x)
  (! (bvzero? (core:bv-bit (ffs x) x)))

to

  (bvzero? (bvand x (bvsub1 (bvshl (bv 1 64) (ffs x)))))
  (! (bvzero? (bvand x (bvshl (bv 1 64) (ffs x)))))

This appears to be slightly easier for the solver.

The second is to revert the change to using Rosette 4's `assume`
to encode the axiomatization. `assume` is sound, but entangles these
assumptions with all later assertions; even unrelated ones. The
solver appears to not like this. Instead, we save these assumptions
in a list and apply them only for the one assertion it matters
for: showing that the final BPF and target registers match.

In the future, could consider writing a library for expressing such
"inline" assumptions as a more general way to avoid use of quantifiers.